requirements.txt
astroid==2.0.4
colorama==0.4.0
Django==2.1.3
django-appconf==1.0.2
django-geoposition==0.3.0
django-measurement==3.1.1
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
measurement==2.0.1
mpmath==1.0.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.7
six==1.11.0
sympy==1.3
wrapt==1.10.11
